#74: Move proto helpers to proto module
