+++ proto/src/main/kotlin/com/commonwealthrobotics/bowlerkernel/proto/ScriptHostProtoUtil.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with bowler-kernel.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.commonwealthrobotics.bowlerkernel.testutil
+package com.commonwealthrobotics.bowlerkernel.proto
 
 import com.commonwealthrobotics.proto.gitfs.FileSpec
 import com.commonwealthrobotics.proto.gitfs.ProjectSpec
