+++ scripting/src/test/kotlin/com/commonwealthrobotics/bowlerkernel/scripting/DefaultScriptTest.kt
@@ -16,8 +16,8 @@
  */
 package com.commonwealthrobotics.bowlerkernel.scripting
 
-import com.commonwealthrobotics.bowlerkernel.testutil.fileSpec
-import com.commonwealthrobotics.bowlerkernel.testutil.projectSpec
+import com.commonwealthrobotics.bowlerkernel.proto.fileSpec
+import com.commonwealthrobotics.bowlerkernel.proto.projectSpec
 import io.kotest.assertions.arrow.either.shouldBeLeft
 import io.kotest.assertions.arrow.either.shouldBeRight
 import io.kotest.assertions.throwables.shouldThrow
