+++ testUtil/build.gradle.kts
@@ -1,8 +1,6 @@
 description = "Test utilities for internal use."
 
 dependencies {
-    api(project(":proto"))
-
     api(group = "org.junit.jupiter", name = "junit-jupiter-api", version = Versions.junit)
     api(group = "org.junit.jupiter", name = "junit-jupiter-engine", version = Versions.junit)
     api(group = "org.junit.jupiter", name = "junit-jupiter-params", version = Versions.junit)
@@ -17,8 +15,4 @@ dependencies {
     api(group = "io.arrow-kt", name = "arrow-core-data", version = Versions.arrow)
 
     api(group = "org.jacoco", name = "org.jacoco.agent", version = Versions.jacocoTool)
-}
-
-tasks.named("classes").configure {
-    dependsOn(project(":proto").tasks.named("generateProto"))
 }
